+++ encapsia_cli/encapsia.py
@@ -85,8 +85,14 @@ def get_user_config():
     "--host",
     help="Name to use to lookup credentials in .encapsia/credentials.toml",
 )
+@click.option(
+    "--silent",
+    is_flag=True,
+    default=False,
+    help="Suppress normal output.",
+)
 @click.pass_context
-def main(ctx, colour, host):
+def main(ctx, colour, host, silent):
     """CLI to talk to an encapsia host.
 
     Options can be provided in one of three ways, in this priority order:
@@ -115,7 +121,7 @@ def main(ctx, colour, host):
 
     """
     ctx.color = {"always": True, "never": False, "auto": None}[colour]
-    ctx.obj = dict(host=host)
+    ctx.obj = dict(host=host, silent=silent)
 
 
 COMMANDS = [