+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased] - 
+
+### Fixed
+
+### Added
+
+- Global --silent flag, for suppressing normal output. #55.
+
 ## [0.5.3] - 2023-01-10
 
 ### Fixed